+++ src/devices/hmIPDevices/hmIpRoll.ts
@@ -23,6 +23,8 @@ export class HmIpRoll extends HmIPDevice implements iShutter {
   public targetAutomaticValue: number = 0;
   /** @inheritDoc */
   public blockAutomationHandler: BlockAutomaticHandler;
+  /** @inheritDoc */
+  public baseAutomaticLevel: number = 0;
   private _setLevelSwitchID: string;
   private _setLevel: number = -1;
   private _setLevelTime: number = -1;